+++ apps/src/code-studio/activityUtils.js
